# src/routes/pagamentos.routes.js
const express = require('express');
const router = express.Router();
const db = require('../config/database');

// POST /api/pagamentos/:consulta_id — chama CALL sp_registrar_pagamento
router.post('/:consulta_id', async (req, res) => {
  try {
    const { consulta_id } = req.params;
    const { forma_pagamento } = req.body;

    if (!forma_pagamento) {
      return res.status(400).json({ erro: 'Campo obrigatório: forma_pagamento' });
    }

    const [rows] = await db.query(
      'CALL sp_registrar_pagamento(?, ?)',
      [consulta_id, forma_pagamento]
    );

    res.json(rows[0][0]);
  } catch (error) {
    console.error('Erro ao registrar pagamento:', error.message);

    if (error.sqlState === '45000') {
      return res.status(422).json({ erro: error.message });
    }
    res.status(500).json({ erro: 'Erro ao registrar pagamento', detalhes: error.message });
  }
});

// GET /api/pagamentos
router.get('/', async (req, res) => {
  try {
    const [rows] = await db.query('SELECT * FROM pagamentos');
    res.json(rows);
  } catch (error) {
    res.status(500).json({ erro: 'Erro ao listar pagamentos', detalhes: error.message });
  }
});

// GET /api/pagamentos/:id
router.get('/:id', async (req, res) => {
  try {
    const [rows] = await db.query('SELECT * FROM pagamentos WHERE id = ?', [req.params.id]);
    if (rows.length === 0) return res.status(404).json({ erro: 'Pagamento não encontrado' });
    res.json(rows[0]);
  } catch (error) {
    res.status(500).json({ erro: 'Erro ao buscar pagamento', detalhes: error.message });
  }
});

// PUT /api/pagamentos/:id
router.put('/:id', async (req, res) => {
  try {
    const { valor_pago, forma_pagamento, status } = req.body;
    const [result] = await db.query(
      'UPDATE pagamentos SET valor_pago=?, forma_pagamento=?, status=? WHERE id=?',
      [valor_pago, forma_pagamento, status, req.params.id]
    );
    if (result.affectedRows === 0) return res.status(404).json({ erro: 'Pagamento não encontrado' });
    res.json({ mensagem: 'Pagamento atualizado com sucesso' });
  } catch (error) {
    res.status(500).json({ erro: 'Erro ao atualizar pagamento', detalhes: error.message });
  }
});

// DELETE /api/pagamentos/:id
router.delete('/:id', async (req, res) => {
  try {
    const [result] = await db.query('DELETE FROM pagamentos WHERE id = ?', [req.params.id]);
    if (result.affectedRows === 0) return res.status(404).json({ erro: 'Pagamento não encontrado' });
    res.json({ mensagem: 'Pagamento deletado com sucesso' });
  } catch (error) {
    res.status(500).json({ erro: 'Erro ao deletar pagamento', detalhes: error.message });
  }
});

module.exports = router;